Example: Use http handler for writing to the console
